+++ assets/js/admin-builder-post-submissions.js
@@ -1,3 +1,5 @@
+/* global wpforms_builder */
+
 'use strict';
 
 ( function( $ ) {
@@ -33,17 +35,12 @@
 		 */
 		bindUIActions: function() {
 
-			// When a featured image field is configured, configure that file
-			// upload field to only accept images.
-			$( document ).on( 'change', '#wpforms-panel-field-settings-post_submissions_featured', function() {
-
-				var fieldID = $( this ).find( 'option:selected' ).val();
-
-				if ( fieldID !== '' ) {
-					$( '#wpforms-field-option-' + fieldID + '-extensions' ).val( 'jpg,jpeg,png,gif' );
-					$( '#wpforms-field-option-' + fieldID + '-max_file_number' ).val( 1 );
-				}
-			} );
+			$( document )
+				.on( 'wpformsBuilderSetupReady', WPFormsPostSubmissions.togglePostSubmissions )
+				.on( 'input', '#wpforms-panel-field-settings-post_submissions', WPFormsPostSubmissions.togglePostSubmissions )
+				.on( 'input', '#wpforms-panel-field-settings-post_submissions_featured', WPFormsPostSubmissions.changeFeaturedImage )
+				.on( 'click', '.wpforms-post-submissions-disabled-option', WPFormsPostSubmissions.disabledMediaAlert )
+				.on( 'wpformsFieldDuplicated', WPFormsPostSubmissions.togglePostSubmissions );
 		},
 
 		/**
@@ -74,6 +71,87 @@
 					},
 				},
 				effect: 'appear',
+			} );
+		},
+
+		/**
+		 * Adjust featured image settings on builder load.
+		 *
+		 * @since 1.5.0
+		 */
+		togglePostSubmissions: function() {
+
+			const disabledClass = 'wpforms-post-submissions-disabled-option';
+
+			$( '.' + disabledClass ).removeClass( disabledClass );
+
+			if ( ! $( '#wpforms-panel-field-settings-post_submissions' ).is( ':checked' ) ) {
+				return;
+			}
+
+			const fieldId = $( '#wpforms-panel-field-settings-post_submissions_featured' ).val();
+			const $options = $( '#wpforms-field-option-' + fieldId );
+
+			$options
+				.find( '.wpforms-field-option-row-extensions' ).addClass( disabledClass )
+				.find( 'input' ).val(  wpforms_builder.post_submissions.image_extensions );
+			$options
+				.find( '.wpforms-field-option-row-max_file_number' ).addClass( disabledClass )
+				.find( 'input' ).val( 1 );
+			$options
+				.find( '.wpforms-field-option-row-media_library' ).addClass( disabledClass )
+				.find( 'input' ).prop( 'checked', 'checked' ).trigger( 'input' );
+		},
+
+		/**
+		 * Adjust featured image settings on featured image field change.
+		 *
+		 * @since 1.5.0
+		 */
+		changeFeaturedImage: function() {
+
+			const fieldId = $( this ).find( 'option:selected' ).val();
+
+			WPFormsPostSubmissions.togglePostSubmissions();
+
+			if ( ! fieldId ) {
+				return;
+			}
+
+			$.confirm( {
+				title: wpforms_builder.heads_up,
+				content: wpforms_builder.post_submissions.disabling_options,
+				icon: 'fa fa-exclamation-circle',
+				type: 'orange',
+				buttons: {
+					confirm: {
+						text: wpforms_builder.ok,
+						btnClass: 'btn-confirm',
+						keys: [ 'enter' ],
+					},
+				},
+			} );
+		},
+
+		/**
+		 * If user clicks on disabled media field, show alert.
+		 *
+		 * @since 1.5.0
+		 */
+		disabledMediaAlert: function() {
+
+			$.alert( {
+				title: wpforms_builder.heads_up,
+				content: wpforms_builder.post_submissions.disabled_option,
+				icon: 'fa fa-exclamation-circle',
+				type: 'orange',
+				buttons: {
+					confirm: {
+						text: wpforms_builder.ok,
+						btnClass: 'btn-confirm',
+						keys: [ 'enter' ],
+					},
+				},
 			} );
 		},
 	};